updating to work with 64bit monomac compile

+++ MonoMac.Windows.Forms/Extenders/FontExtender.cs
@@ -22,7 +22,7 @@ namespace System.Windows.Forms
 		{
 			if (font == null)
 				return new System.Drawing.Font ("Arial", 9.9f, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
-			return new System.Drawing.Font (font.FontName, font.PointSize, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+			return new System.Drawing.Font (font.FontName, (float)font.PointSize, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
 			
 		}
 		public static NSFont ToNsFont (this Font font)
